refactor runner configuration
